--- br/pkg/redact/BUILD.bazel
@@ -1,25 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "redact",
-    srcs = ["redact.go"],
-    importpath = "github.com/pingcap/tidb/br/pkg/redact",
-    visibility = ["//visibility:public"],
-    deps = ["@com_github_pingcap_errors//:errors"],
-)
-
-go_test(
-    name = "redact_test",
-    timeout = "short",
-    srcs = [
-        "main_test.go",
-        "redact_test.go",
-    ],
-    embed = [":redact"],
-    flaky = True,
-    deps = [
-        "//pkg/testkit/testsetup",
-        "@com_github_stretchr_testify//require",
-        "@org_uber_go_goleak//:goleak",
-    ],
-)
