+++ pkg/util/redact/redact.go
@@ -17,6 +17,7 @@ package redact
 import (
 	"bufio"
 	"bytes"
+	"encoding/hex"
 	"fmt"
 	"io"
 	"os"
@@ -171,4 +172,35 @@ func DeRedact(remove bool, input io.Reader, output io.Writer, sep string) error 
 	}
 
 	return nil
+}
+
+// InitRedact inits the enableRedactLog
+func InitRedact(redactLog bool) {
+	mode := errors.RedactLogDisable
+	if redactLog {
+		mode = errors.RedactLogEnable
+	}
+	errors.RedactLogEnabled.Store(mode)
+}
+
+// NeedRedact returns whether to redact log
+func NeedRedact() bool {
+	mode := errors.RedactLogEnabled.Load()
+	return mode != errors.RedactLogDisable && mode != ""
+}
+
+// Value receives string argument and return omitted information if redact log enabled
+func Value(arg string) string {
+	if NeedRedact() {
+		return "?"
+	}
+	return arg
+}
+
+// Key receives a key return omitted information if redact log enabled
+func Key(key []byte) string {
+	if NeedRedact() {
+		return "?"
+	}
+	return strings.ToUpper(hex.EncodeToString(key))
 }
